refactor: split FeatureRequestModal into focused modules
Split the 2049-line FeatureRequestModal.tsx into 6 focused files:
- FeatureRequestTypes.ts (125 lines) - types, constants, utilities
- FeedbackDialogs.tsx (311 lines) - discard confirm, login prompt, preview overlays
- DraftsTab.tsx (172 lines) - drafts list tab
- UpdatesTab.tsx (870 lines) - requests + GitHub contributions tab
- SubmitTab.tsx (728 lines) - submit form + success view + footer
- FeatureRequestModal.tsx (433 lines) - slim entry point

Partially addresses #8624
